+++ git-remote-updater.php
@@ -1,29 +1,29 @@
 <?php
 /**
- * Git Bulk Updater
+ * Git Remote Updater
  *
  * @author  Andy Fragen
  * @license MIT
- * @link    https://github.com/afragen/git-bulk-updater
- * @package git-bulk-updater
+ * @link    https://github.com/afragen/git-remote-updater
+ * @package git-remote-updater
  */
 
 /**
- * Plugin Name:       Git Bulk Updater
- * Plugin URI:        https://github.com/afragen/git-bulk-updater
+ * Plugin Name:       Git Remote Updater
+ * Plugin URI:        https://github.com/afragen/git-remote-updater
  * Description:       Allows you to easily update GitHub Updater repositories in bulk via RESTful endpoint updating.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           0.2.5
+ * Version:           0.3.0
  * License:           MIT
  * Domain Path:       /languages
- * Text Domain:       git-bulk-updater
- * GitHub Plugin URI: https://github.com/afragen/git-bulk-updater
+ * Text Domain:       git-remote-updater
+ * GitHub Plugin URI: https://github.com/afragen/git-remote-updater
  * Requires PHP:      7.1
  * Requires WP:       5.1
  */
 
-namespace Fragen\Git_Bulk_Updater;
+namespace Fragen\Git_Remote_Updater;
 
 /*
  * Exit if called directly.
@@ -34,5 +34,5 @@ if ( ! defined( 'WPINC' ) ) {
 }
 
 // Setup plugin loading.
-require_once __DIR__ . '/src/Git_Bulk_Updater/Bootstrap.php';
+require_once __DIR__ . '/src/Git_Remote_Updater/Bootstrap.php';
 ( new Bootstrap( __FILE__ ) )->run();