+++ src/Git_Remote_Updater/Actions.php
@@ -1,14 +1,14 @@
 <?php
 /**
- * Git Bulk Updater
+ * Git Remote Updater
  *
  * @author  Andy Fragen
  * @license MIT
- * @link    https://github.com/afragen/git-bulk-updater
- * @package git-bulk-updater
+ * @link    https://github.com/afragen/git-remote-updater
+ * @package git-remote-updater
  */
 
-namespace Fragen\Git_Bulk_Updater;
+namespace Fragen\Git_Remote_Updater;
 
 /*
  * Exit if called directly.
@@ -39,7 +39,7 @@ class Actions {
 	 */
 	public function add_plugin_menu() {
 		global $_registered_pages;
-		if ( isset( $_registered_pages['settings_page_git-bulk-updater'] ) ) {
+		if ( isset( $_registered_pages['settings_page_git-remote-updater'] ) ) {
 			return;
 		}
 
@@ -48,10 +48,10 @@ class Actions {
 
 		add_submenu_page(
 			$parent,
-			esc_html__( 'Git Bulk Updater', 'git-bulk-updater' ),
-			esc_html__( 'Git Bulk Updater', 'git-bulk-updater' ),
+			esc_html__( 'Git Remote Updater', 'git-remote-updater' ),
+			esc_html__( 'Git Remote Updater', 'git-remote-updater' ),
 			$capability,
-			'git-bulk-updater',
+			'git-remote-updater',
 			[ $this, 'create_admin_page' ]
 		);
 	}
@@ -60,10 +60,10 @@ class Actions {
 	 * Options page callback.
 	 */
 	public function create_admin_page() {
-		$action = is_multisite() ? 'edit.php?action=git-bulk-updater' : 'options.php';
+		$action = is_multisite() ? 'edit.php?action=git-remote-updater' : 'options.php';
 
 		echo '<div class="wrap"><h2>';
-		esc_html_e( 'Git Bulk Updater', 'git-bulk-updater' );
+		esc_html_e( 'Git Remote Updater', 'git-remote-updater' );
 		echo '</h2>';
 
 		$this->show_feedback();
@@ -72,11 +72,11 @@ class Actions {
 		echo '<form method="post" action="' . esc_attr( $action ) . '">';
 		echo '<table class="form-table">';
 
-		echo '<tbody class="git-bulk-updater-repo">';
+		echo '<tbody class="git-remote-updater-repo">';
 		( new Actions_Row() )->add_repo_rows();
 		echo '</tbody>';
 
-		echo '<tbody class="git-bulk-updater-site">';
+		echo '<tbody class="git-remote-updater-site">';
 		( new Actions_Row() )->add_site_rows();
 		echo '</tbody>';
 
@@ -90,10 +90,10 @@ class Actions {
 	 * @return void
 	 */
 	private function show_feedback() {
-		$feedback = get_site_transient( 'git_bulk_updater_feedback' );
+		$feedback = get_site_transient( 'git_remote_updater_feedback' );
 		if ( $feedback ) {
 			echo '<div>';
-			echo '<h3>' . esc_html__( 'Update Feedback', 'git-bulk-updater' ) . '</h3>';
+			echo '<h3>' . esc_html__( 'Update Feedback', 'git-remote-updater' ) . '</h3>';
 			foreach ( $feedback as $repo_feedback ) {
 					echo '<div><p>';
 				foreach ( $repo_feedback as $message ) {
@@ -110,12 +110,12 @@ class Actions {
 	 */
 	private function repo_or_site_selector() {
 		$options = [
-			'git-bulk-updater-repo' => esc_html__( 'Show Repositories', 'git-bulk-updater' ),
-			'git-bulk-updater-site' => esc_html__( 'Show Sites', 'git-bulk-updater' ),
+			'git-remote-updater-repo' => esc_html__( 'Show Repositories', 'git-remote-updater' ),
+			'git-remote-updater-site' => esc_html__( 'Show Sites', 'git-remote-updater' ),
 		];
 		?>
-		<label for="git-bulk-updater">
-			<select id="git-bulk-updater" name="git-bulk-updater">
+		<label for="git-remote-updater">
+			<select id="git-remote-updater" name="git-remote-updater">
 				<?php foreach ( $options as $key => $value ) : ?>
 						<option value="<?php esc_attr_e( $key ); ?>" <?php selected( $key ); ?> >
 							<?php esc_html_e( $value ); ?>
@@ -134,7 +134,7 @@ class Actions {
 	public function redirect() {
 		$redirect_url = is_multisite() ? network_admin_url( 'settings.php' ) : admin_url( 'tools.php' );
 		$location     = add_query_arg(
-			[ 'page' => 'git-bulk-updater' ],
+			[ 'page' => 'git-remote-updater' ],
 			$redirect_url
 		);
 		wp_safe_redirect( $location );
@@ -147,12 +147,12 @@ class Actions {
 	 * @return void
 	 */
 	public function load_js() {
-		if ( isset( $_GET['page'] ) && 'git-bulk-updater' === $_GET['page'] ) {
+		if ( isset( $_GET['page'] ) && 'git-remote-updater' === $_GET['page'] ) {
 			add_action(
 				'admin_enqueue_scripts',
 				function () {
-					wp_register_script( 'git-bulk-updater-actions', plugins_url( basename( GIT_BULK_UPDATER_DIR ) . '/js/git-bulk-updater-switcher.js' ), [], false, true );
-					wp_enqueue_script( 'git-bulk-updater-actions' );
+					wp_register_script( 'git-remote-updater-actions', plugins_url( basename( GIT_REMOTE_UPDATER_DIR ) . '/js/git-remote-updater-switcher.js' ), [], false, true );
+					wp_enqueue_script( 'git-remote-updater-actions' );
 				}
 			);
 		}